minishell tester fixes (wip)
realized why Hubert and Eliz couldn't run the minishell tester that night
because we were using rl_done; so switch back to use other rl functions
and minishell tester is running.

realized in the tester that relative path is needed so i added a protype
function for the relative path.

also realized from the test cases that many logic needs to be changed
especially for the error codes / exit status, particularly in the redirects

still fixing and testing in progress and valgrind check is not done yet and
norm check not done yet too.